+++ HISTORY.rst
@@ -14,11 +14,13 @@ Unreleased
 * ALCYN-1761: the "Related" button on the job related to invoices now try to open the correct view (customer or supplier)
 * Take only pickings to generate the delivery round
 * Delivery round: Cancel BO for customer that do not want BO
+* ALCYN-1563: Import the final customer balance and post all new journal entries
 
 **Bugfixes**
 
 * ALCYN-1767: Fix delivery note csv for product with no vat info on sale order, look for vat in product directly.
 * ALCYN-1762: fix transactional errors during background delivery of delivery rounds
+* Fix missing records during a reception with additional product
 
 **Build**